agents/azure-ai-foundry-a2a/requirements.txt
uvicorn[standard]
a2a-sdk
httpx
click
langchain
langchain-core
langchain-openai
langgraph
python-dotenv
pydantic